requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boto3=1.24.2=py310h06a4308_0
botocore=1.27.2=py310h06a4308_0
bottleneck=1.3.4=py310h9102076_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.4.26=h06a4308_0
certifi=2022.6.15=py310h06a4308_0
cffi=1.15.0=py310hd667e15_1
cryptography=37.0.1=py310h9ce1e76_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
freetype=2.10.4=h0708190_1
giflib=5.2.1=h36c2ea0_2
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py310h06a4308_0
ipython=8.3.0=py310h06a4308_0
jedi=0.18.1=py310h06a4308_1
jmespath=0.10.0=pyhd3eb1b0_0
jpeg=9e=h166bdaf_1
jupyter_client=7.2.2=py310h06a4308_0
jupyter_core=4.10.0=py310h06a4308_0
kiwisolver=1.4.2=py310h295c915_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=11.2.0=h1234567_1
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h9c3ff4c_1
matplotlib-base=3.4.3=py310h23f4a51_2
matplotlib-inline=0.1.2=pyhd3eb1b0_2
missingno=0.4.2=py_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py310h06a4308_0
numexpr=2.8.1=py310hcea2de6_2
numpy=1.22.3=py310hfa59a62_0
numpy-base=1.22.3=py310h9585f30_0
openssl=1.1.1o=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py310h295c915_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.2=pyhd8ed1ab_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=21.2.4=py310h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2022.1=py310h06a4308_0
pyzmq=22.3.0=py310h295c915_2
readline=8.1.2=h7f8727e_1
s3transfer=0.6.0=py310h06a4308_0
scipy=1.7.3=py310hfa59a62_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
setuptools=61.2.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.2=py310h96516ba_0
tk=8.6.12=h1ccaba5_0
tornado=6.1=py310h7f8727e_0
traitlets=5.1.1=pyhd3eb1b0_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zeromq=4.3.4=h2531618_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=ha95c52a_0
